refactor(search): Search, performSearch, CacheAndUpdateResults
