+++ src/search/search.go
@@ -1,139 +1,43 @@
 package search
 
 import (
-	"context"
-	"fmt"
-	"net/url"
-	"strings"
-	"time"
-
 	"github.com/hearchco/hearchco/src/anonymize"
-	"github.com/hearchco/hearchco/src/bucket"
 	"github.com/hearchco/hearchco/src/bucket/result"
-	"github.com/hearchco/hearchco/src/category"
+	"github.com/hearchco/hearchco/src/cache"
 	"github.com/hearchco/hearchco/src/config"
 	"github.com/hearchco/hearchco/src/engines"
-	"github.com/hearchco/hearchco/src/rank"
 	"github.com/rs/zerolog/log"
-	"github.com/sourcegraph/conc"
 )
 
-func PerformSearch(query string, options engines.Options, conf *config.Config) []result.Result {
-	searchTimer := time.Now()
-
-	relay := bucket.Relay{
-		ResultMap: make(map[string]*result.Result),
+func Search(query string, options engines.Options, conf *config.Config, db cache.DB) ([]result.Result, bool) {
+	var results []result.Result
+	var foundInDB bool
+	gerr := db.Get(query, &results)
+	if gerr != nil {
+		// Error in reading cache is not returned, just logged
+		log.Error().
+			Err(gerr).
+			Str("queryAnon", anonymize.String(query)).
+			Str("queryHash", anonymize.HashToSHA256B64(query)).
+			Msg("cli.Run(): failed accessing cache")
+	} else if results != nil {
+		foundInDB = true
+	} else {
+		foundInDB = false
 	}
 
-	query, timings, toRun := procBang(query, &options, conf)
-
-	query = url.QueryEscape(query)
-	log.Debug().
-		Str("queryAnon", anonymize.String(query)).
-		Str("queryHash", anonymize.HashToSHA256B64(query)).
-		Msg("Searching")
-
-	resTimer := time.Now()
-	log.Debug().Msg("Waiting for results from engines...")
-	var worker conc.WaitGroup
-	runEngines(toRun, timings, conf.Settings, query, &worker, &relay, options)
-	worker.Wait()
-	log.Debug().
-		Int64("ms", time.Since(resTimer).Milliseconds()).
-		Msg("Got results")
-
-	rankTimer := time.Now()
-	log.Debug().Msg("Ranking...")
-	results := rank.Rank(relay.ResultMap, conf.Categories[options.Category].Ranking) // have to make copy, since its a map value
-	rankTimeSince := time.Since(rankTimer)
-	log.Debug().
-		Int64("ms", rankTimeSince.Milliseconds()).
-		Int64("ns", rankTimeSince.Nanoseconds()).
-		Msg("Finished ranking")
-
-	log.Debug().
-		Int64("ms", time.Since(searchTimer).Milliseconds()).
-		Msg("Found results")
-
-	return results
-}
-
-// engine_searcher, NewEngineStarter()  use this.
-type EngineSearch func(context.Context, string, *bucket.Relay, engines.Options, config.Settings, config.Timings) error
-
-func runEngines(engs []engines.Name, timings config.Timings, settings map[engines.Name]config.Settings, query string, worker *conc.WaitGroup, relay *bucket.Relay, options engines.Options) {
-	config.EnabledEngines = engs
-	log.Info().
-		Int("number", len(config.EnabledEngines)).
-		Str("engines", fmt.Sprintf("%v", config.EnabledEngines)).
-		Msg("Enabled engines")
-
-	engineStarter := NewEngineStarter()
-	for i := range engs {
-		eng := engs[i] // dont change for to `for _, eng := range engs {`, eng retains the same address throughout the whole loop
-		worker.Go(func() {
-			// if an error can be handled inside, it wont be returned
-			// runs the Search function in the engine package
-			err := engineStarter[eng](context.Background(), query, relay, options, settings[eng], timings)
-			if err != nil {
-				log.Error().
-					Err(err).
-					Str("engine", eng.String()).
-					Msg("search.runEngines(): error while searching")
-			}
-		})
-	}
-}
-
-func procBang(query string, options *engines.Options, conf *config.Config) (string, config.Timings, []engines.Name) {
-	useSpec, specEng := procSpecificEngine(query, options, conf)
-	goodCat := procCategory(query, options)
-	if !goodCat && !useSpec && query[0] == '!' {
-		// options.category is set to GENERAL
+	if foundInDB {
 		log.Debug().
 			Str("queryAnon", anonymize.String(query)).
 			Str("queryHash", anonymize.HashToSHA256B64(query)).
-			Msg("search.procBang(): invalid bang (not category or engine shortcut)")
-	}
-
-	query = trimBang(query)
-
-	if useSpec {
-		return query, conf.Categories[category.GENERAL].Timings, []engines.Name{specEng}
+			Msg("Found results in cache")
 	} else {
-		return query, conf.Categories[options.Category].Timings, conf.Categories[options.Category].Engines
-	}
-}
-
-func trimBang(query string) string {
-	if (query)[0] == '!' {
-		return strings.SplitN(query, " ", 2)[1]
-	}
-	return query
-}
-
-func procSpecificEngine(query string, options *engines.Options, conf *config.Config) (bool, engines.Name) {
-	if query[0] != '!' {
-		return false, engines.UNDEFINED
-	}
-	sp := strings.SplitN(query, " ", 2)
-	bangWord := sp[0][1:]
-	for key, val := range conf.Settings {
-		if strings.EqualFold(bangWord, val.Shortcut) || strings.EqualFold(bangWord, key.String()) {
-			return true, key
-		}
+		log.Debug().
+			Str("queryAnon", anonymize.String(query)).
+			Str("queryHash", anonymize.HashToSHA256B64(query)).
+			Msg("Nothing found in cache, doing a clean search")
+		results = performSearch(query, options, conf)
 	}
 
-	return false, engines.UNDEFINED
-}
-
-func procCategory(query string, options *engines.Options) bool {
-	cat := category.FromQuery(query)
-	if cat != "" {
-		options.Category = cat
-	}
-	if options.Category == "" {
-		options.Category = category.GENERAL
-	}
-	return cat != ""
+	return results, foundInDB
 }
